Marcador por nota e inscrito

+++ surtus-api-restful/Dtos/Responses/Inscripcion/DatosMarcadorInscritoResponse.cs
@@ -4,6 +4,6 @@ namespace surtus_api_restful.Dtos.Responses.Inscripcion
     public class DatosMarcadorInscritoResponse
     {
         public string Marcador { get; set; }
-        public double Nota { get; set; }
+        public decimal Nota { get; set; }
     }
 }
